chore: moving destroy to BaseAnnotator too
