+++ src/document/DocumentAnnotator.ts
@@ -42,10 +42,6 @@ export default class DocumentAnnotator extends BaseAnnotator {
             this.highlightListener.destroy();
         }
 
-        if (this.deselectManager) {
-            this.deselectManager.destroy();
-        }
-
         if (this.managers) {
             this.managers.forEach(managers => managers.forEach(manager => manager.destroy()));
         }
