+++ src/document/__tests__/DocumentAnnotator-test.ts
@@ -112,17 +112,6 @@ describe('DocumentAnnotator', () => {
 
             expect(mockManager.destroy).toHaveBeenCalled();
         });
-
-        test('should destroy DeselectManager', () => {
-            annotator = getAnnotator();
-
-            const mockDeselectManager = ({ destroy: jest.fn() } as unknown) as DeselectManager;
-            annotator.deselectManager = mockDeselectManager;
-
-            annotator.destroy();
-
-            expect(mockDeselectManager.destroy).toHaveBeenCalled();
-        });
     });
 
     describe('event handlers', () => {
